Changing default file format name
git-svn-id: svn+ssh://sxs-archive.tapir.caltech.edu/SpEC/SVNroot/trunk/Support/WaveExtrapolation/Waveforms@17722 aeca8a61-ec13-0410-a8d4-8d6e48c7bcfb

+++ Utilities/VectorFunctions.cpp
@@ -1125,7 +1125,7 @@ vector<double> WU::Unwrap(const vector<double>& In) {
   return Out;
 }
 
-vector<double>& WU::Unwrap(vector<double>& Vec, const double i1, const double i2) {
+vector<double>& WU::Unwrap(vector<double>& Vec, const unsigned int i1, const unsigned int i2) {
   // Compare Matlab's unwrap.m file
   double Dp = 0.0;
   double Dps = 0.0;
